refactor: Fallback 컴포넌트 및 스타일 분리
