+++ src/components/Fallback/Fallback.style.ts
@@ -1,0 +1,17 @@
+import { css } from "@emotion/react";
+export const fallbackContainerStyle = css({
+  display: "flex",
+  flexDirection: "column",
+  alignItems: "center",
+  gap: "4rem",
+  height: "800px",
+  width: "320px",
+});
+
+export const fallbackImageStyle = css({
+  width: "400px",
+});
+
+export const fallbackTitleStyle = css({
+  fontWeight: "bold",
+});
